# A script for managing sparse checkouts