--- a/notes
+++ b/notes
@@ -1,0 +1,3 @@
+vscode startup: 48s vs 1m06s
+
+cargo check after modifying one file: 45s vs 2m16s