By using sparse checkouts, you can make rust-analyzer much faster, since it doesn't need to process every crate in the repo.

In a decent sized diff, in which the following crates were checked out

        crates/sui-node
        crates/sui-single-node-benchmark
        crates/sui-benchmark
        crates/sui-core
        crates/sui-types
        crates/sui
        crates/mysten-common

rust-analyzer for vscode was much faster both for startup and for running cargo check after a single file was edited.

Startup speed (until cargo check stopped on its own):
- before: 1m6s at best, some runs took many minutes for some reason
- after: 50s on average

Cargo check after editing a single file, no errors:
- before: 2m16s
- after 45s

cargo check, time from introducing a simple error to finishing
- before 12s
- after 7s

Of course, the editor is responsive before these processes fully complete, but faster
is still better than slower

This also makes it far faster to run various commands from the root of the repo such as
`cargo check`, `clippy`, `nextest run` and so on. These can often be sped up with `-p` or `--test` as well,
but that can be tricky to use correctly in some cases, and its very convenient to be able to quickly
run only the tests in crates you've edited.